rename yeelight to yeelightblue
Related to: https://github.com/eclipse/smarthome/pull/3531#issuecomment-304609859
