+++ extensions/binding/org.eclipse.smarthome.binding.yeelightblue/.classpath
