+++ extensions/binding/org.eclipse.smarthome.binding.yeelightblue/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>org.eclipse.smarthome.binding.yeelight</name>
+	<name>org.eclipse.smarthome.binding.yeelightblue</name>
 	<comment></comment>
 	<projects>
 	</projects>
