+++ extensions/binding/org.eclipse.smarthome.binding.yeelightblue/META-INF/MANIFEST.MF
@@ -1,9 +1,9 @@
 Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
-Bundle-Name: Bluetooth Low Energy Binding
-Bundle-SymbolicName: org.eclipse.smarthome.binding.ble;singleton:=true
+Bundle-Name: Yeelight Blue (BLE) Binding
+Bundle-SymbolicName: org.eclipse.smarthome.binding.yeelightblue;singleton:=true
 Bundle-Vendor: Eclipse.org/SmartHome
-Bundle-Version: 0.8.0.qualifier
+Bundle-Version: 0.9.0.qualifier
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-ClassPath: .
 Import-Package: 
@@ -18,9 +18,7 @@ Import-Package:
  org.eclipse.smarthome.io.transport.bluetooth,
  org.eclipse.smarthome.io.transport.bluetooth.events,
  org.eclipse.smarthome.io.transport.bluetooth.le,
- org.osgi.framework;version="1.8.0",
- org.osgi.service.component;version="1.2.2",
+ org.osgi.framework,
+ org.osgi.service.component,
  org.slf4j
 Service-Component: OSGI-INF/*.xml
-Export-Package: org.eclipse.smarthome.binding.yeelight,
- org.eclipse.smarthome.binding.yeelight.handler
