+++ extensions/binding/org.eclipse.smarthome.binding.yeelightblue/OSGI-INF/YeelightDiscovery.xml
@@ -8,9 +8,9 @@
     http://www.eclipse.org/legal/epl-v10.html
 
 -->
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.eclipse.smarthome.binding.yeelight.discovery">
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.eclipse.smarthome.binding.yeelightblue.discovery">
 
-   <implementation class="org.eclipse.smarthome.binding.yeelight.internal.discovery.YeelightDiscoveryService"/>
+   <implementation class="org.eclipse.smarthome.binding.yeelightblue.internal.discovery.YeelightDiscoveryService"/>
 
    <service>
       <provide interface="org.eclipse.smarthome.config.discovery.DiscoveryService"/>
