+++ extensions/binding/org.eclipse.smarthome.binding.yeelightblue/OSGI-INF/YeelightHandlerFactory.xml
@@ -8,9 +8,9 @@
     http://www.eclipse.org/legal/epl-v10.html
 
 -->
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.eclipse.smarthome.binding.yeelight.internal.YeelightHandlerFactory">
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.eclipse.smarthome.binding.yeelightblue.internal.YeelightHandlerFactory">
 
-   <implementation class="org.eclipse.smarthome.binding.yeelight.internal.YeelightHandlerFactory"/>
+   <implementation class="org.eclipse.smarthome.binding.yeelightblue.internal.YeelightHandlerFactory"/>
 
    <service>
       <provide interface="org.eclipse.smarthome.core.thing.binding.ThingHandlerFactory"/>
