+++ extensions/binding/org.eclipse.smarthome.binding.yeelightblue/build.properties
