+++ extensions/binding/pom.xml
@@ -30,7 +30,7 @@
     <module>org.eclipse.smarthome.binding.wemo</module>
     <module>org.eclipse.smarthome.binding.wemo.test</module>
     <module>org.eclipse.smarthome.binding.yahooweather</module>
-    <module>org.eclipse.smarthome.binding.yeelight</module>
+    <module>org.eclipse.smarthome.binding.yeelightblue</module>
   </modules>
 
 </project>
